Tell første utfall metrikker samtidig som alle utfall.
